requirements.txt
praw
yt-dlp
moviepy
pillow
pyyaml 